Move tab generation to its own class
